requirements.txt
fastapi
uvicorn[standard]
Jinja2
python-multipart
httpx
python-dotenv
email-validator
